Prevent to use `.group` in V2::SearchesAPI
Solve MySQL GROUP() issue. SQLite3 is ok but MySQL is not.
Mysql2::Error: Expression #1 of SELECT list is not in GROUP BY clause and contains nonaggregated column 'amis_new_development.terms.id' which is not functionally dependent on columns in GROUP BY clause; this is incompatible with sql_mode=only_full_group_by